% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{arable}
\alias{arable}
\title{arable API client}
\value{
An object of class \code{arable}.
}
\description{
handles the authentication scheme and provides a basic ping method.
}
\examples{
\dontrun{
ar <- arable$new(username = Sys.getenv("AR_USERNAME"),
  password = Sys.getenv("AR_PASSWORD"))
ar
ar$get_auth()
ar$ping()

arable$new(apikey = Sys.getenv("AR_APIKEY"))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{root url to api}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{arable$print()}}
\item \href{#method-new}{\code{arable$new()}}
\item \href{#method-get_auth}{\code{arable$get_auth()}}
\item \href{#method-ping}{\code{arable$ping()}}
\item \href{#method-clone}{\code{arable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for \code{client}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{arable$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}

\item{\code{x}}{self}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new client object using different methods.
The Arable API provides three methods of authentication to the API:
username + password (basic auth), apikey or bearer token.
Precedance is token > apikey > username + password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{arable$new(username, password, apikey, token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{username}

\item{\code{password}}{password}

\item{\code{apikey}}{apikey}

\item{\code{token}}{token}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_auth"></a>}}
\if{latex}{\out{\hypertarget{method-get_auth}{}}}
\subsection{Method \code{get_auth()}}{
Get authentication header
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{arable$get_auth()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a httr request object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ping"></a>}}
\if{latex}{\out{\hypertarget{method-ping}{}}}
\subsection{Method \code{ping()}}{
Ping API by calling /devices endpoint.
Returns TRUE if we get a 200
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{arable$ping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{arable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
