% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{arable_options}
\alias{arable_options}
\title{Set or get options for my package}
\usage{
arable_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{base_url}}{ The base url for Arable API}
 \item{\code{email}}{ The email for Arable}
 \item{\code{password}}{ The password for Arable}
 \item{\code{tenant}}{ The tenant for Arable}
 \item{\code{apikey}}{ The apikey for Arable}
}
}

