% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arable.R
\name{connect}
\alias{connect}
\title{Connect to the Arable API}
\usage{
connect(email = Sys.getenv("ARABLE_EMAIL"),
  password = Sys.getenv("ARABLE_PASSWORD"),
  tenant = Sys.getenv("ARALE_TENANT"), apikey = Sys.getenv("ARABLE_APIKEY"))
}
\arguments{
\item{email}{The email address}

\item{password}{The password}

\item{tenant}{The company name}

\item{apikey}{The API key (optional)}
}
\value{
TRUE if connected
}
\description{
Connect to the Arable API
}
