% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{ar_devices}
\alias{ar_devices}
\alias{ar_parse_devices}
\title{List all devices}
\usage{
ar_devices(ar, page = 1L, limit = 24L, verbose = FALSE)

ar_parse_devices(x)
}
\arguments{
\item{ar}{A \code{\link{arable}} object. Required.}

\item{page}{integer; starting page of api to receive data}

\item{limit}{integer; number of records per page (default: 24, max 10000)}

\item{verbose}{logical; verbose messages on API calls}

\item{x}{a devices object, as returned by \code{\link[=ar_devices]{ar_devices()}}}
}
\description{
List all devices

parse devices into a nice data.frame
}
\examples{
\dontrun{
ar <- arable$new(username = Sys.getenv("AR_USERNAME"),
  password = Sys.getenv("AR_PASSWORD"))
devices <- ar_devices(ar, limit = 100L)
ar_parse_devices(devices)
}
}
