% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arable.R
\name{get_data}
\alias{get_data}
\title{Get data from Arable for devices}
\usage{
get_data(devices, start, end, location = NULL, format = "csv",
  order = "time", measure = "hourly", select = "all", limit = 1000)
}
\arguments{
\item{devices}{A character vector of device names}

\item{start}{The start timestamp}

\item{end}{The end timestamp}

\item{location}{optional; id of a location to retrieve data for; devices ignored if this is present}

\item{format}{optional; use format=csv to get csv-formatted data; otherwise data is returned as json}

\item{order}{optional; "time" (time ascending) or "-time" (time descending)}

\item{measure}{optional; "calibrated", "hourly", or "daily"}

\item{select}{optional; "all", "spectrometer", "microclimate", or "radiometer"}

\item{limit}{optional; maximum number of data points to return; defaults to 1000}
}
\value{
A data frame of data for devices
}
\description{
Get data from Arable for devices
}
