% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arable.R
\name{get_devices}
\alias{get_devices}
\title{List information of all devices in Arable}
\usage{
get_devices(device_id = NULL, name = NULL)
}
\arguments{
\item{device_id}{The ID of device}

\item{name}{The name of device}
}
\value{
A list of device information
}
\description{
List information of all devices in Arable
}
