% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{ar_schemas}
\alias{ar_schemas}
\alias{ar_schema}
\title{List all schemas}
\usage{
ar_schemas(ar, verbose = FALSE)

ar_schema(ar, table, verbose = FALSE)
}
\arguments{
\item{ar}{A \code{\link{arable}} object. Required.}

\item{verbose}{logical; verbose messages on API calls}

\item{table}{tablename, as returned by \code{\link[=ar_schemas]{ar_schemas()}}}
}
\description{
List all schemas

Show schema of a table
}
\examples{
\dontrun{
ar <- arable$new(username = Sys.getenv("AR_USERNAME"),
  password = Sys.getenv("AR_PASSWORD"))
ar_schemas(ar)
}
\dontrun{
ar <- arable$new(username = Sys.getenv("AR_USERNAME"),
  password = Sys.getenv("AR_PASSWORD"))
ar_schema(ar, "daily")
ar_schema(ar, "hourly")
}
}
